Add 2020-10
For part A, counting the jumps was pretty trivial. For part B, I began
with an attempt to generate all the possible paths using recursion. This
worked well for the smaller samples, but with the full input this proved
to be woefully inefficient.

It took me so long fumbling about for a good strategy. Eventually I
landed on the idea to track the sums of paths using a multidimensional
array. I tracked the valid connecting adaptors for every adaptor
(regardless if those adaptors existed or not). Creating the sum in this
way results in the total pathways counted for the final adaptor.