Add 2020-12
This one was fun, prompting me to draw coordinates on paper to wrap my
head around the puzzle. Fortunately for part A, I chose to represent
directions as a pair of coordinate deltas. It turned out that part B
introduced the "waypoint", which is pretty much the same thing, albeit
the waypoints change in more interesting ways.

I particularly liked being able to use Python's multiple assignment
feature to rotate coordinates like:

  x, y = y, -x