Add 2020-15
Solving for part A was fairly easy. I put the starting numbers into an
list, and just kept appending to the list following the rules of the
game. But to implement those rules, I did counts in the list to identify
"new" vs "old" numbers. I also used a list compression within a max() to
identify the highest previous index. All of that proved to be far too
slow for part B. It was never going to finish.

To solve part B, I changed my program to use a dictionary to track the
history of the highest previous index. This removed all of the
unnecessary searching. Calculating both answers takes just under
10 seconds. There's likely a much faster way.