Add 2020-17
The first part of this puzzle took me longer than necessary because I
didn't read the rules carefully. I mistakenly had my code deactivating
cubes with 2 or 3 neighbors.

The aspect of this puzzle I found tricky was understanding the need to
start with a larger space than what is described in the input. There are
cubes outside of the input-described cubes that are affected during the
first cycle.

For part B, it was not too difficult to change my neighbor searching
code to work for a dynamic set of dimensions.