Add 2020-18
This was a mix of looping and regex. For part A, I used a regex to
identify inner groupings, and used recursion to replace the inner
groupings with their calculated results.

Then for part B, it was easy to use the same strategy, but match
individual additions, processing all of those before the rest of the
expression.