Add 2020-21
My strategy for this was to count the frequency of ingredients for
each allergen. Then identify each allergen when an allergen had a single
high-frequency ingredient that was not an already identified allergen.

In the end, what took an embarrassing long time to debug was in the code
which parsed the input. I used re.findall() to parse the input, but just
ended the regex with '\n'. However, this left out the last line. :(

This didn't cause a problem for the sample input. Once I added a '?' to
the end of the regex, I was able to get the correct answers.