Add 2020-22
Part A was pretty easy. I checked ahead of time to verify there were no
overlapping cards, and then used deques to represent the decks.

Part B took a longer than necessary because I didn't understand the
rules. I solved part B for the sample data, but it did not solve for the
puzzle input. It turned out I didn't understand what was meant by
"instant end" rule. I thought it referred to the cards "played" in the
round, but it really meant the state of the decks before the round
begins. Once I fixed that I got the correct answer.