Add 2020-23
I solved part A using a deque. That was far too inefficient for part B.
It actually took quite a bit of re-reading for me to understand how the
puzzle wanted the set of cups expanded. Something about how it is worded
was confusing to me.

What eventually worked for part B was doing away with all of the popping
and inserting, and instead maintaining a doubly linked list. First I
looked for a Python library, and found llist.dllist. I looked at this a
bit, but it seems very much like deque, and there's no way to directly
change the next and prev items. So I ended up creating a Cup class to
track the next and prev cups.