Add 2020-24
This was more interesting than the other coordinate system puzzles. I
started by drawing some hexagons and writing in coordinates. I used a
three-coordinate tuple, x = e-w, y = nw-se, z = ne-sw. I picked a hex to
be 0, 0, 0 and worked out the vectors for the adjacent tiles. Then I
just added the vectors to origin to locate the tiles that needed to be
flipped.

For part B, a key important step was expanding the starting tiles
(exactly like the Day 17 puzzle).